Pruebas y Lint corregidos

+++ eslint.config.js
@@ -6,6 +6,6 @@ export default {
   },
   rules: {
     semi: ['error', 'always'],
-    quotes: ['error', 'double'],
+    quotes: ['error', 'single'],
   },
 };
